Merge 182719 - Fix paste not showing up on empty text fields
If we move the handle the paste popup hides.  So we were trying to move the handle to the
right position after showing the paste popup.  Now we show/move the handle first, then show
the paste popup.  There is still a bug with the auto zoom hiding the paste popup but that
requires a larger change that moves all handle positions to document space.

BUG=176193


Review URL: https://chromiumcodereview.appspot.com/12253043

TBR=dtrainor@chromium.org
Review URL: https://codereview.chromium.org/12300028