[android_webview] Fix race conditions in AwContentsTest.
Waiting for onPageFinished is not a sufficient condition for a different
AwContentsClient callback to fire. Due to the Java memory model even if a
callback is guaranteed to happen at a later time than onPageFinished there is
no guarantee that the other thread will see the updates.
I also used this as an opportunity to split stuff out from AwContentsTest
because it was way too big.
Finally, I've re-enabled some of the tests since they've been passing locally for me.

BUG=None
TEST=AndroidWebViewTests

Review URL: https://codereview.chromium.org/12279004