Removes the arrow from the Windows notification center, aligns to nearest edge.
Corrects a bug where the alignment was getting set to the side when the taskbar was auto-hide.  New behavior is that the alignment is only left or right-oriented when the taskbar is fixed on the left or right of the screen.

Corrects a bug in the calculation of the maximum height of the message center.

TEST=double-check ChromeOS TrayBubbleViews, manual inspection of Windows.
BUG=174417

Review URL: https://chromiumcodereview.appspot.com/12254017