Temporarily disable InterProcessTimeTicksConverter.
This is a speculative fix for bug 174170. Google's metrics show a discrepancy
in PLT in M24. That was when we fully enabled IPTTC. I'm going to disable it for
a couple of days and see if their metrics return to normal. If so, we'll have to
do some further debugging.


BUG=174170


Review URL: https://chromiumcodereview.appspot.com/12271006