Drop down options are disabled for a bookmark in the bookmark manager
In r172230 there is an acknowledgement that the active element for the canExecute calls cannot be relied on and rather the use of the global |list| should be used to get the selected items in the bookmark list. However, |updateOpenCommands| still used the |e.target| which is not referencing the selected node.

This wasn't a problem until r181240 where the |canExecute| in |MenuButton.showMenu| now calls |MenuButton.updateCommands| which eventually calls |updateOpenCommands| with |e.target| set to the button rather than a |list| element, thus causing the problem.

The fix involves using the |list.selectedItems| instead of |e.target.selectedItem| in |updateOpenCommands|.

After this issue is resolved there is an additional issue that is showing up where clicking to bring up the menu then clicking the menu item will close the menu entirely before the onMouseUp is handled. I will report this as a separate issue and work on a patch.

Contributed by: tom.cassiotis@gmail.com

TEST=Various selections of combinations of folders and bookmarks (nothing, single bookmark, multiple bookmarks, single folder with and without bookmarks inside it, combination bookmark and folder etc).

BUG=174562
TBR=tom.cassiotis@gmail.com

Review URL: https://codereview.chromium.org/12317005