cc: Don't sync tilings to layers that don't draw content.
If a layer is drawsContent == false, it should have no tilings, which
we assert to be the case. When the pending tree is created, tilings are
synced back to the pending tree, but not for layers that do not draw
content in the pending tree. However, if we create a tiling for the
active tree later, we sync that tiling to the pending tree even if the
layer does not draw content, creating a situation where a non-content-
drawing layer has tilings but it should not.

Similarly, if the pending tree creates a tiling, but the layer already
exists in the active tree in a state that does not draw content, the
tiling should not be added to the active layer.

R=reveman,enne
BUG=176524
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/12291002