Reset the date/time picker client in the RenderViewImpl destructor.
Because the RenderDateTimePicker class is a RenderViewObserver, it'll
register itself on the RenderViewImpl class causing it to be destroyed
as the very last thing in ~RenderViewImpl() user code.  However, because
it's stored in a scoped_ptr<> we can end up deleting it twice.

BUG=176495
NOTRY=True

Review URL: https://chromiumcodereview.appspot.com/12298022