Apply the pref that prohibits deleting browsing and download history.
When kAllowDeletingBrowserHistory is false, the browsing-history and download items are disabled and unchecked in the Clear Browsing Data UI, the handler removes them from the requested data types to remove if the UI is bypassed, and the BrowsingDataRemover ignores them if they're requested. Belt, suspenders, and duct tape, anyone?

Note that this does not mean that history items will stick around forever. Users will still be able to delete or edit the underlying databases in the file system directly (except on Chrome OS), and Chrome itself may decide to expire or archive history data at any time.

BUG=175419
TEST=new browser and unit tests; also verify that "clear history" and "clear downloads" checkboxes in the Clear Browsing Data dialog are unchecked and disabled if kAllowDeletingBrowserHistory is false.

Review URL: https://chromiumcodereview.appspot.com/12262008