History: Eliminate query cursor, use end_time argument instead.
The query cursor was introduced in http://crrev.com/171498 to allow
querying for results a page at a time. For history sync, it's
necessary to truncate the list of local results when they are merged
with the results from the server. When results are truncated, the
cursor becomes useless. This CL removes the cursor in favor of an
end_time argument that specifies the end of the query range.

BUG=176812


Review URL: https://chromiumcodereview.appspot.com/12218147