Android WebView: update directories that are removed.
Add a number of directories to the known issue list for the Android
WebView merge process; these are not directories with license problems,
but rather directories that contain code we want to be certain we are
not using to build. Previously we had some issues where we were
inadvertantly including Chromium's in-tree headers for libraries that we
linked to the system version of, causing ABI compatibility problems.

BUG=

Review URL: https://codereview.chromium.org/12255071