tryjob parser used to accept try jobs formatted like builder,builder:test:filter,test:filter,builder
where builder is specified after a test/filter.  This is ambiguous because
if we don't have a full set of builder names, then a ',' after a filter
could either a test or a builder.

Also now it drops builders that it does not recognize.

BUG=

Review URL: https://chromiumcodereview.appspot.com/12226027