Use _target_name instead of package_name for java*.gypi
We were using package_name as a unique name for naming output files and
directories. package_name was typically the same as _target_name or a
variation of it (like dropping _apk). Using _target_name instead means
we need to specify one less thing and it is (maybe?) guaranteed to be
unique.

TBR=brettw,joi,jar,fischman,zea,sky
BUG=

Review URL: https://chromiumcodereview.appspot.com/11308030