IndexedDB: Restore maximum value size checks and tests
The plumbing to limit IDB values to 64MB to avoid hitting IPC limitations was removed during various refactors and the tests were disabled. This restores the check and tests. The failure is now reported to script as an asynchronous error callback rather than as a synchronous exception.

BUG=160577,120298
R=alecflett@chromium.org,dgrogan@chromium.org


Review URL: https://chromiumcodereview.appspot.com/12317023