Store the content bounds of the shell window, rather than the frame bounds.
Before, windows would gain $TITLEBAR_HEIGHT pixels in height each time they
were closed and opened again.

Also, make the system remember the restored bounds of the window instead of the
current bounds (i.e. if the window is maximized when it is closed, then when it
is reopened it will have the bounds it had before it was maximized.)

BUG=160343


Review URL: https://chromiumcodereview.appspot.com/12282022