Android buildbot - halt build if subcommand returns error
bb_run_bot is currently ORing the return codes of each
subcommand but not early terminating.  That means that if
a command returns an error, it does not explicitly mark
that step as failing. The quickest fix is to early terminate,
which is correct for the current usage of bb_run_bot, where
subcommands do not return non-zero status except on
serious error.

Example problematic build:
http://build.chromium.org/p/chromium.linux/builders/Android%20Tests/builds/1479
Review URL: https://codereview.chromium.org/12324007