touch: Change the behaviour of compositor-thread hittesting for touch events.
The ACK state of each touch-point is maintained individually. So, if a touch
point has no consumer for the touch-start event, then no subsequent move or
end events for that point will be dispatched to the renderer. However, during
this touch sequence, if any new touch-point is pressed, then the press event
will be dispatched to the renderer. Whether subsequent move/end events for
the second touch-point are dispatched to the renderer depends entirely on the
ACK state for the press-event of the second touch-point, and is completely
independent of the ACK state of the first touch point.

BUG=174762

Review URL: https://codereview.chromium.org/12320004