Expose FS show/hide thresholds to Command Line
Pipe the show/hide thresholds out to the command line.  Allow for them not to be
based on one parameter (there isn't one split point that determines show/hide).
Instead there are two thresholds, one for showing and one for hiding.  The
ambiguous area is handled by determining whether or not the scroll was, as a
whole, moving up or down.

BUG=174449
NOTRY=true


Review URL: https://chromiumcodereview.appspot.com/12210050