cc: Added more items into RasterTask(Meta)Data
This change renames RasterTaskMetadata, and adds more items into
the struct. This is done to reduce the number of parameters passed
to PerformRaster, and to allow adding more flags to be passed,
without hitting base::Bind limit of 7 arguments.

testing: Manually tested with 1, 2, and 15 raster threads on
theverge.com and cnn.com


Review URL: https://chromiumcodereview.appspot.com/12252054