Add an app list view delegate for the Cocoa app list, and add basic keyboard usage.
This change adds a concrete type for AppListControllerDelegate to allow 
the ui code to communicate back to the browser code. In order for the 
app list to be dismissed, a key handler is added to the first responder 
such that pressing escape will dismiss the app list via the delegate. 
The key handler also handles basic keyboard navigation and selection 
highlighting.

BUG=138633
TEST=Covered by app_list_unitests, and browser_tests: AppListControllerBrowserTest.ShowAndDismiss

Review URL: https://chromiumcodereview.appspot.com/12294007