Try a few times to let the initial added URLs show up in the history database.
An unknown file-locking or asynchronous operation sometimes prevents the two added URLs from showing up immediately in the history database, leading to *more* history results after the (prohibited, no-op) deletion than before it. We'll try a few times, waiting a little in between each (10 tries, 100ms each try, still giving a total of only 1s delay), to see if that eliminates the flakiness.

The test is, of course, fundamentally still flaky on Windows, as it must be if writing to the history database is itself unreliable. If the flakiness continues to show up in practice, the test will have to be disabled on Windows.

BUG=176828
TEST=HistoryApiTest.DeleteProhibited no longer flaky on Windows (XP, Win7, Vista)

Review URL: https://chromiumcodereview.appspot.com/12319013