Installing ChromeOS font on developer Linux machine
Added call to a new script in:
build/install-build-deps.sh 
to download and install the ChromeOS default font
(Noto Sans UI). The installation is in:
/usr/local/share/fonts.
ChromeOS on Linux will use the font if present.
The installation creates (or recreates) a subdir:
"chromeos". In it it creates a .stamp file which is used
to avoid re-installation if it is already up to date.
It also creates a README file.

BUG=148945
TEST=manual;
One way to test is to instrument the file:
src/ui/gfx/platform_font_pango.cc
to log default_font_->GetFontName(); in ctor
PlatformFontPango::PlatformFontPango()
Expect:
W/out font installation: DejaVu Sans
With font installation: Noto Sans UI



Review URL: https://chromiumcodereview.appspot.com/12313039