Don't unpin the normal Chrome shortcut when deleting profile shortcuts.
Running the unpin code inside ShellUtil::RemoveShortcut() caused the
unpinning of a regular Chrome shortcut even when the shortcut being
deleted was a profile shortcut.

Fix this by using file_util::Delete() directly, rather than going
through ShellUtil::RemoveShortcut(), which does a lot of unnecessary
things, including unpinning the shortcut, which is not desired here.

I checked that none of the other things RemoveShortcut() does is
needed in this case (for example, it compares the name of executable
in the shortcut with the specified one, but that's already handled in
this codepath).

Unfortunately, this can be tested with unit tests because the taskbar
state cannot be mocked and tests shouldn't be affecting global
Windows state.

Note: This will likely need some revisiting once we support properly
pinning profile shortcuts.

BUG=176642
TEST=As described in bug.

Review URL: https://chromiumcodereview.appspot.com/12330026