Fix crash in VideoRendererBase::ThreadMain().
The interleaving of VideoRendererBase::Stop() with outstanding asynchronous calls to a VideoDecoder would result in continuing to execute code (e.g., starting up a thread) on a stopped VideoRendererBase. The most common manifestation of the bug was crashing in ThreadMain() as VideoRendererBase was being destroyed.

The CHECK()s added to ~VideoRendererBase() in r180573 will stay there as they've proven useful.

BUG=174150


Review URL: https://chromiumcodereview.appspot.com/12324005