cc: Make ResourceProvider correctly handle transferring mailbox-created resources
For mailbox-created resources, the texture ID is created lazily, in lockForRead. If we use a delegating renderer, we won't lock it, so at the time we transfer it to the parent and back, we don't have a texture ID, and we shouldn't be producing/consuming. We still need to correctly handle the synchronization logic.

This change implements all this, and adds tests.

BUG=164095


Review URL: https://chromiumcodereview.appspot.com/12315046