cc: Give mask and replica layers a parent.
Mask and replica layers hang off of another layer in the tree. They are
not strictly children of the layer but they are attached to it. This
patch lets them find the layer they are hanging off, and calls that
the layer's parent. We then call them a "dependent" of the layer rather
than a "child" to distinguish them but still denote a bidirectional
relationship.

This will allow us to early out for entire subtrees in the push-properties
tree walk (and potentially other tree walks as well), to greatly speed up
commit.

Tested by LayerTest.maskAndReplicaHasParent

BUG=177756
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/12334046