Add support for a manager passphrase of a managed account.
The managed user settings page will be shown with every control disabled. A click on the button "Unlock" shows the passphrase dialog. After a successful authentication, the controls of the settings page are enabled. The authentication is reset when the settings page
is closed.

BUG=171370
TEST=browser_tests

Review URL: https://chromiumcodereview.appspot.com/11783008