content: Introduce WebContentsViewPort.
WebContentsImpl calls into WebContentsView when a new render-view gets created
(via RenderViewCreated). When navigating in history, a render-view can become
active without having to be recreated first. So introduce a callback that gets
triggered in these cases. Instead of adding to the public interface in
WebContentsView, introduce WebContentsViewPort that all platform implementations
inherit from, and add the callback to that interface.

Use this new interface to make sure the navigation overlay is observing the
correct RenderWidgetHost for paint updates.

BUG=177554

Review URL: https://codereview.chromium.org/12320058