Added PrinterJobQueueHandler class and tests
Added handler class and tests for the printer job
queue. This class will eventually allow the connector
to select the most appropriate job in the case of failure
to fetch the print ticket or file data for one or more jobs
on the print queue, and includes facilities for
exponential backoff.

Removed redundaunt data structures in PrinterJobHandler,
and replaced implementation of HandlerJobMetadata response
with one that uses PrinterJobQueueHandler but only acts if there is
no wait time for a job.

This CL does not change the actual behavior of the connector
at the moment, but introduces code that will be used to
allow the CL to more intelligently select jobs in case of a
fetch failure in the future.


BUG=


Review URL: https://chromiumcodereview.appspot.com/12260012