Force alternative path for ShowLoginPopup to use the active desktop to create the login popup.
I couldn't find any case where "current_login_ui()" isn't already set. This seems outdated now that we use a WebUI for signin. I couldn't strip it out either however and in an effort to remove CreateParams methods without HostDesktopType this has to change...

The issue with multi-desktop and singleton dialogs is obviously that if they are activated on one desktop from another it isn't visible to the user... the sign-in entries however appear to be grayed out while a signin is in progress so this is not such an issue here (maybe we'll want to allow a signin dialog to be active on each desktop eventually, but for now this is a simple step towards multi-desktop aware code).

BUG=129187


Review URL: https://chromiumcodereview.appspot.com/12315037