Extend the CloudPolicyValidator to support ExternalPolicyData.
The CloudPolicyValidator verifies that policy downloaded from the cloud server
is valid and checks its signature. This CL specializes the template to parse
ExternalPolicyData protobufs, which are used for extensions policy.

NOTE: This is a reland of http://crrev.com/184157 without modifications; it was reverted
because DeviceLocalAccountTest.PolicyDownload failed, but that test is flaky
and fails for unrelated reasons (http://crbug.com/177880).

BUG=163318
TBR=mnissler@chromium.org


Review URL: https://chromiumcodereview.appspot.com/12320079