Move WideToASCII to the base namespace.
I accidentally changed some callers to use "base::" for this call. Since the
plan is to move this to the base namespace eventually, I went ahead and added
it with a "using" (to keep the non-converted code compiling) rather than revert
my other changes.

TBR=sky
Review URL: https://codereview.chromium.org/12330103