x11 desktop aura: Fix the cached bounds of the X window.
The bounds in ConfigureNotify events are in the parent window's coordinate
system, and the parent window may not be the X root window. So translate
the origin of the window to the X root window's coordinate system first.
This fixes the positioning of menus under some window managers.

BUG=none

Review URL: https://codereview.chromium.org/12319091