IndexedDB: Ensure cursor prefetch count grows as expected
Cursor prefetching was supposed to exponentially increase the number of records
fetched each time, but had a subtle bug so that kept the amount fixed, limiting
performance - mihaip@chromium.org found the bug and identified the fix.

Also added a unit test for prefetching to verify that prefetching is initiated
after some threshold of continue() calls, and that the prefetch count
increases over time.

BUG=167520,118033


Review URL: https://chromiumcodereview.appspot.com/12313072