Ensure that flash fullscreen mode works in Chrome desktop AURA.
Currently fullscreen flash only works in Chrome ASH.

When fullscreen mode is requested by a plugin (flash) a new RenderWidgetHostViewAura
instance is created and its RenderWidgetHostViewAura::InitAsFullscreen method is invoked.

In Chrome ASH on Windows 8 and Chrome OS we find a root window with matching bounds as
those passed in and also eventually resize the aura::Window instance maintained by the
RenderWidgetHostViewAura created above after reparenting it to the RootWindow found above.

In desktop Chrome we create a top level fullscreen widget and set it as the parent of the
aura::Window instance hosted by the RenderWidgetHostViewAura class.

The code to create a temporary top level fullscreen widget is provided by a new class
DesktopNativeWidgetFullscreenHandler which is created as a singleton. This class watches the
RenderWidgetHostViewAura window and when it is destroyed destroys the top level fullscreen widget.

I also fixed the DCHECK which was occurring in the RenderWidgetHostViewAura::TransientWindowObserver
class on Windows due to setting an observer on an aura::Window instance which was already
setup in the contructor of this class. The DCHECK fires because of attempting to add the 
same observer twice.

Fixes bugs https://code.google.com/p/chromium/issues/detail?id=174753 and 
http://code.google.com/p/chromium/issues/detail?id=177478

BUG=174753,177478
TEST=Covered by existing ppapi browser test FlashFullscreen which was disabled originally
for AURA and two new tests
1. WidgetTest.DesktopAuraFullscreenChildParentDestroyed 
2. WidgetTest.DesktopAuraFullscreenChildDestroyedBeforeParentTest
Review URL: https://codereview.chromium.org/12320046