Add infrastructure for AnnotatedFactory
The CL puts in place infrastructure for a fully-annotated factory. This
consists of a factory class that serves the same purpose as, say,
GClientFactory; a commands class to provide various forms of annotated commands
to the factory (currently only an "execute this script" command is implemented);
and support scripts in scripts/slave to actually execute the specified script.

It also portends a change in information location. In particular, the
factory_properties that used to be specified in master.cfg can now be specified
in "recipe" files (generally found in scripts/slave/recipes). This allows
factory_properties changes to be made on the slave without requiring a master
restart.


Review URL: https://chromiumcodereview.appspot.com/12320028