Remove SyncSessionJob::DidReachServer() flag
This flag was added in r140885 to address issue 131414.  The problem was
that SyncSessionJobs that did not contact the server would succeed and
attempt to reset the backoff interval. The commits r149052 and r143088
removed the job types that do not attempt to contact the server, so we
no longer need to worry about them accidentally distrupting backoff.

BUG=175024


Review URL: https://chromiumcodereview.appspot.com/12320104