Split WebContents and navigation notifications out of GoogleURLTracker
Separates the content-implementation-specific aspects of
GoogleURLTracker--WebContents usage, and notifications from the
implementation of NavigationController--into a helper class. This class
also hides the extra complexity of Instant, so that it is transparent
to GoogleURLTracker.

This is primarily to support GoogleURLTracker use in iOS, which doesn't
use WebContents or NavigationControllerImpl. An iOS implementation
of GoogleURLTrackerNavigationObserver will follow at some later point.

BUG=169452

Review URL: https://codereview.chromium.org/11856011