Bluetooth: Ignore the disconnect error message when device is already disconnected.
In some cases, we may show an error message while trying to disconnect
a disconnected device. In those cases we should not even be able to
call Disconnect(), but if we do (for example calling Disconnect() twice)
don't show the error message to the user.

BUG=chromium-os:34418
TEST=Reprocude the steps in the bug. No error message while trying to disconnect must be shown.


Review URL: https://chromiumcodereview.appspot.com/12315067