ChromeFrame sites in IE would appear truncated at times.
This was a regression caused by the change to invalidate the cached protocol data when IE (urlmon)
terminates the protocol. While this works well for non CF requests, for CF requests when we report
the changed mime type to be that of chrome, the protocol is terminated with cached data. When Chrome
attempts to read it we don't have anything to report which leads to the problem.

Fix is to not invalidate the protocol data for Chrome requests in our Terminate hook. We should invalidate
it for Abort because that seems like an error condition.

BUG=178421
R=robertshield.
Review URL: https://chromiumcodereview.appspot.com/12310145