IndexedDB: Re-enable key-type-array test as FLAKY to check current behavior
This test started failing flakily several months ago; the root cause was
not identified but appeared to be a difficult to repro v8/binding/gc issue.
Unable to repro locally but some extensive work has been done in this
area recently, so we want to enable the test to see if it's still flaking.

BUG=165671

Review URL: https://codereview.chromium.org/12334099