Initialize BrowserPolicyConnector after the IOThread is created.
The import process on Windows spawns a subprocess and spins a nested message loop waiting for it to complete. If this subprocess takes longer than 5 seconds then the message loop will execute a delayed initialization task of the policy code, that grabs the system request context; however, since the IOThread hasn't been created yet this leads to a crash.

BUG=177843

Review URL: https://chromiumcodereview.appspot.com/12310100