Do not drop video frame that is about to be rendered.
In the previous CL (r183625), we drop a video frame when it's late. However, we
check that condition after sleep and when we are about to render that frame. If
everything works correctly, the remaining time is 0. Given some timer error,
which fluctuates around 0, we end up dropping about half of all frames! 

This CL fixes this issue and takes the timing error into account. If a frame is
late because the timing fluctuation, it'll not be dropped. But if a frame is
really late (e.g. due to decoding delay), it will still be dropped. 

BUG=178619,chrome-os-partner:18000,chrome-os-partner:18003
TEST=Tested on Daisy. The number of dropped frames is much less.
Review URL: https://codereview.chromium.org/12314126