Note avatar icon is shown mirrored in tab strip on Mac/Retina
Avatar icons for Mac is mirrored because they appear at the other end.
But note icon is special because it has polarity. It should not be mirrored.

In 100% images, Chrome already use the same image for the both platform.
But, in 200% images, the icon is mistakenly mirrored.

BUG=178397
TEST=manual

Review URL: https://codereview.chromium.org/12313149