Make SimpleEntryImpl::Close asynchronous.
Since only one entry operation can be in flight at once, it's safe to assume that if the IO thread Entry doesn't exist on completion that the sync entry should be deleted.

This issue is downstream of https://codereview.chromium.org/12223075/ and should land after it.

R=rvargas@chromium.org
BUG=173392

Review URL: https://chromiumcodereview.appspot.com/12277004