Improve unloading/loading launchd jobs in Chromoting Mac installer
This fixes the installer in the case where there are multiple
users logged in, or the installer is being run from a
switched-out session. Previously, if there were multiple
instances of the loginwindow process running, only one of them
(randomly selected) would be used for stopping or starting the
host service.
With this CL, the preflight script records the users with host
processes running, and the postflight script restarts the host
process for each of those users.

BUG=176355


Review URL: https://chromiumcodereview.appspot.com/12319105