For user-resizing, panel should update its bounds before calling OnPanelResizedByMouse.
SetPanelBoundsInstantly should not be called within PanelCollection::OnPanelResizedByMouse. Instead, it should be done in the resize controller before calling OnPanelResizedByMouse. This is because panels on Windows rely on system resizing, instead of resize controller. When the Windows system notifies about the bounds change, we should not try to set bounds again. Without this fix, the logic to update custom max size on user resizing will be omitted on Windows.

BUG=none
TEST=existing tests


Review URL: https://chromiumcodereview.appspot.com/12326149