[win] Clicking on the taskbar icon should hide the launcher.
This CL changes the conditions under which the launcher is kept visible
after losing focus. It now loses visibility if the taskbar gets focused
and the right mouse button is not down. This still allows right clicking
on its taskbar icon, for example to pin it, but solves the reported bug.

BUG=178902


Review URL: https://chromiumcodereview.appspot.com/12326154