Show a warning overlay and connection interstitial for out-of-date hosts.
Out-of-date hosts can be identified from the hostVersion field in the directory.
They are identified in the host-list using the existing "warning" icon overlaid
on the host icon. Off-line hosts are never flagged as out-of-date, as this may
not indicate a problem (if they have been off-line for a long time). Attempting
to connect to an out-of-date host shows a warning interstitial explaining that
a manual update is required (and can be done while connected).

The CL also cleans up the HostTableEntry class, better separating the duties of
the ctor, create (now createDom) and init methods.

BUG=177122


Review URL: https://chromiumcodereview.appspot.com/12342024