Split knowledge of WebContents out of OmniboxEditModel.
Separates the elements of OmniboxEditModel that interact with the current page via WebContents into a new OmniboxCurrentPageDelegate.

The goal of this change is to support OmniboxEditModel usage on iOS, which
doesn't use WebContents. An iOS implementation of OmniboxCurrentPageDelegate will follow at a later point.


Review URL: https://chromiumcodereview.appspot.com/12304023