Ignore the disabled state of component-updated PPAPI Flash for once.
In the early stage of component-updated PPAPI Flash, we did field trials in which it was set to disabled by default. The corresponding settings item might remain in some users' profiles. Currently it affects both the bundled and component-updated PPAPI Flash (since the two share the same enable/disable state). We want to remove this item to get those users to use PPAPI Flash.

We will want to do so only once.

BUG=None
TEST=None


Review URL: https://chromiumcodereview.appspot.com/12388012