Fix touch scaling for high dpi windows.
Touch events in windows are always expressed in screen coordinates, even if
the target app is not dpi-aware. This patch determines the dpi scale used
by the OS, and re-scales touch events accordingly. 

In some cases the touch events are sent to the wrong window. This is detected
and corrected.

BUG=175542


Review URL: https://chromiumcodereview.appspot.com/12317157