[content shell] sync the navigation state to the browser before recording the navigation history
In layout tests that dump the back/forward history, the state of the latest
navigation entry (which is kept in the browser browser) does not reliably
reflect the current state. This is because updates to the state are only send
when a new navigation starts or else at regular intervals.

Therefore, we force a sync of the session history state when running layout
tests that dump the back/forward history.

BUG=111316
R=brettw@chromium.org
TEST=remaining http/tests/navigation tests pass

Review URL: https://codereview.chromium.org/12330099