Ensure that the fullscreen preference on apps is honored in desktop Chrome on Windows.
Currently if USE_ASH is defined we force the fullscreen pref to downgrade to a full screen window
which is the desired behavior for ASH/Chrome OS, but not for desktop chrome.

Fixes are to ensure that fullscreen apps in desktop Chrome AURA open in fullscreen mode in the current
tab on the same lines as non AURA desktop chrome.

BUG=176544
R=yoz
TBR=benwells
Review URL: https://codereview.chromium.org/12380045