An even better fix for the tab crasher while audio indicator is on
See bug below, basically the issue happens if you
are:
- playing audio
- capturing like with webrtc
- tab changes icon or caption
- capturing stops

I know it sounds like hard to find case however
hangouts makes it happen.

If you can see both the audio indicator which is an
equalizer and the capture which is a red dot,then
you are in the problem zone.


This time arround I have unittests for the relevant
cases.

BUG=178154
TEST= see bug
Review URL: https://codereview.chromium.org/12377038