Minimize GPU memory usage on memory tests on Mac.
On the Mac, GPU resources are counted against CPU memory usage
sometimes (depending on how the OpenGL driver is feeling). This
results in a massive memory regression.

This patch tells the test that it has only 1 MB of GPU memory
to use. Foregrounded renderers will not be held to this limit
(rather, they will be held to a limit of 64MB), but this will
prevent backgrounded renderers from keeping resources around.

BUG=178531
R=zmo@chromium.org


Review URL: https://chromiumcodereview.appspot.com/12389062