Improve host-list display in case of auth or network error.
There were a couple of underlying problems:

1. The user-visible string for NETWORK_FAILED was very host specific. I changed
it to be more generic and added a new NAT_TRAVERSAL_DISABLED error for that
case. Note that there may be non-NAT-related situations that would cause the
same error condition, so it's not 100% accurate. I've also gotten rid of the
NO_RESPONSE error, which was serving the same purpose as NETWORK_FAILED, but
with less helpful language.   

2. The cached host-list was being cleared in case of errors, so the web-app
was unable to display the name of the local host.

BUG=175241


Review URL: https://chromiumcodereview.appspot.com/12374053