Implement "hole" video frame.
In order to support out-of-band compositing of protected video
content in Chrome, it is necessary to have a transparent hole in
Chrome's HTML rendering so that video playing "beneath" the HTML
elements can be seen through.
    
To accomplish the scenario stated above, define a "hole" video frame
which would clear the region where video would have been composited.
In CC, this is done by sending a transparent black SolidColorDrawQuad
with blending disabled.

Review URL: https://chromiumcodereview.appspot.com/12255032