Implemented AdapterState.
I had to use polling to update the adapter since IOBluetoothHostController is not KVO compliant (i.e. we cannot put callbacks or notifiers to listen to the adapter state changes).

But fortunately we do not have to create a separate thread since all the Bluetooth System APIs in MACOSX are asynchronous.

BUG=135472


Review URL: https://chromiumcodereview.appspot.com/12379054