Fix focus for in-tab navigations.
For in-tab navigations, if the contents is supposed to be focused, we should be
calling SetInitialFocus() instead of directly calling Focus() which blindly focuses
the contents view. This way, WebContents can send focus to the location bar on
load if it wants to.

TESTED=manual, per bugs
BUG=179499,175147


Review URL: https://chromiumcodereview.appspot.com/12390042