cc: Make DelegatedRendererLayerImpl::SetRenderPasses() private.
This method was now public only for tests. SetFrameData() should be
used instead. Implemented a SetFrameDataForRenderPasses() method on
the FakeDelegatedRendererLayerImpl class. This class constructs a
frame data out of a set of render passes, making a bogus set up
TransferableResources for each resource in the render passes, and
gives the whole frame to the real DelegatedRendererLayerImpl.

Tested by all the delegated renderer layer tests.

R=piman
BUG=123444


Review URL: https://chromiumcodereview.appspot.com/12388082