Move Mailbox from cc to gpu, and its traits to gpu/ipc
Mailbox is safer to IPC than coaxing it into a string, because otherwise we need
to validate the size when coming from untrusted sources.

This is to enable passing Mailbox through IPC without having to depend on cc.
Among others I will need it to move Pepper to mailboxes, and pepper can't depend
on cc.

BUG=164095


Review URL: https://chromiumcodereview.appspot.com/12378053