Implement Invalidator::Acknowledge.
We implement this by creating a local queue of entries we've received invalidations for and then immediately acknowledging to Tango. When InvalidationHandlers acknowledge that they've finished processing for an id, we erase their entry from the queue; otherwise, we send reminder invalidations on an exponentially increasing delay.

BUG=124149

Review URL: https://chromiumcodereview.appspot.com/10911084