Don't show newly-enabled hosts as out-of-date.
When enabling remote connections for the local system, the
host was being shown as out-of-date, because the dummy host
entry (created for the new host pending a heartbeat) didn't
have any version number set. This CL sets a version that is
new enough not to trigger the warning.


Review URL: https://chromiumcodereview.appspot.com/12388097