Pass the |curtain_activated| flag to IpcDesktopEnvironmentFactory.
In the case of the multi-process host the ability to support curtain mode is a property of the desktop environment representing the screen. This CL hooks up IpcDesktopEnvironmentFactory to receive the request to activate the curtain. This way the factory class will be able to create a different kind of the desktop environment if the curtain was requested.

Other changes:
  - Added remoting_rdp_session% GYP flag that enables RDP integration on Windows.

BUG=137696


Review URL: https://chromiumcodereview.appspot.com/12379084