Expose global state variables for sampling tracing to WebKit
We are implementing TRACE_EVENT macros for sampling profiling. It works in the following mechanism:

- Chromium defines global state variables for sampling profiling. (i.e. g_trace_state0, g_trace_state1, g_trace_state2 in trace_event.h)

- WebKit gets the addresses of the global state variables at the initialization step. (i.e. EventTracer::initialize())

- WebKit updates the global states by using TRACE_EVENT_SAMPLING_STATE() macros every time WebKit changes its state. (e.g. DOM attribute getters/setters/methods)

- A sampling thread running in Chrome reads the global states periodically and visualizes the profiling results into about://tracing.

In this issue, we make a Chromium side change to implement a WebKit API to get the addresses of the global states.

The WebKit patch is here: https://bugs.webkit.org/show_bug.cgi?id=110932

Review URL: https://chromiumcodereview.appspot.com/12316146