Improved CoreAudioUtil::IsSupported().
This patch adds one more test in CoreAudioUtil::IsSupported() and that is to be able to create an IMMDeviceEnumerator. We have seen some crashes where it does not seem to be sufficient to check that Audioses.dll can be loaded. I hope we are fine now.

BUG=166397
TEST=media_unittests and different audio clients in Chrome.

Review URL: https://chromiumcodereview.appspot.com/12378066