Temporary fix for build break when removing Java files or resources.
This appends the list of inputs to the ant and process_resources.py
commands. Thus when the list of inputs changes, the command line
changes, and ninja will re-run the command.

This will be removed once ninja is updated to automatically rebuild
when the input list changes.

BUG=177449,177552


Review URL: https://chromiumcodereview.appspot.com/12379066