[Media Gallery, Win] Implement asynchronous MtpDeviceDelegateImplWin.
Previously,
(1) MTPDeviceDelegateImplWin implemented a synchronous version of MTPDeviceDelegate.
(2) MTPDeviceDelegateImplWin was operated on multiple threads (blocking pool thread and UI thread).
(3) Because of the synchronous behavior, copying a file off an MTP device was blocking the entire thread.

Now,
(1) MTPDeviceDelegateImplWin implements an asynchronous version of MTPDeviceDelegate.
(2) MTPDeviceDelegateImplWin runs on the single thread (IO thread).
(3) MTPDeviceDelegateImplWin supports weak pointers to cancel the callbacks that may run after the destruction of MTPDeviceDelegateImplWin.
(3) Because of the asynchronous behavior, MTPDeviceDelegateImplWin does a call-and-reply to a blocking pool thread to complete the requested device operation (such as GetFileInfo, ReadDirectory, CreateSnapshotFile, etc).
(4) Copying a file off an MTP device is divided into several sub tasks giving a chance to abort the copy operation when required. We no longer block the entire thread to copy the complete MTP device file.

BUG=151679
TEST=none
TBR=kinuko@chromium.org

Review URL: https://chromiumcodereview.appspot.com/12209061