Browser Plugin: Implement BrowserPluginGuestManager
This patch centralizes instance ID allocation and routing to a per-profile object called BrowserPluginGuestManager.

This patch minimizes BrowserPluginEmbedder's responsibilities without entirely removing it.

BrowserPluginEmbedder lives for the lifetime of an embedder WebContents. It handles IPC messages that are not directed at any particular guest process, currently those are:

BrowserPluginHostMsg_AllocateInstanceID
BrowserPluginHostMsg_CreateGuest,
BrowserPluginHostMsg_PluginAtPositionResponse

These message handlers were kept in BrowserPluginEmbedder because they either operate on the embedder (as with PluginAtPosition), or they need to reply back to the correct BrowserPluginManager. This requires an understanding of the  process ID and routing ID of the origin of the message. This information does not seem particularly appropriate in BrowserPluginGuestManager because that object spans multiple embedders and potentially multiple processes (outside of Chrome Apps).

BUG=174673, 140316
Test=BrowserPluginHostTest.*

Review URL: https://chromiumcodereview.appspot.com/12253046