Android: Remove Surface cruft
This removes unused Surface-related Java-level IPC plumbing.
What is not needed anymore is the code to pass Surfaces to child
processes. This was only used in a deprecated render path and for
MediaPlayer being instantiated in the renderer (which is not supported
anymore, it lives in the browser process now).


Review URL: https://chromiumcodereview.appspot.com/12388038