Switch all contexts in a share group to be on the same screen as the window they present to.
When OS X switches GPUs, some contexts that were created beforehand and aren't associated with an NSView can be left behind on the old GPU, which will cause problems when other contexts attempt to share textures with them.

To fix this, track what renderer is being used to present to the screen, transmit that to the IOSurfaceImageTransportSurface in the GPU process, and change every context in the same share group to use that renderer.

BUG=174149


Review URL: https://chromiumcodereview.appspot.com/12383042