Android: use gcc (not cpp) to preprocess templates.
When building the WebView on macs, cpp is a wrapper script around llvm
which does not accept all the same arguments as the cpp that's part of
gcc. Invoke gcc -E instead, which should be guaranteed to accept gcc
arguments (even if it ultimately invokes llvm as well).

BUG=

Review URL: https://codereview.chromium.org/12513008