Windows: BrowserChildProcessHostImpl reports OnProcessCrashed if the child process crashes before connecting the IPC channel.
When a child process crashes, this is generally detected by an error on the IPC channel, i.e. OnChannelError being called. However, if the child process exits or crashes before this IPC channel has been established, the crash is lost. The browser process sees the child process being launched but never learns of its termination.

This patch uses a WaitableEventWatcher to track the lifetime of the child process up to the point the IPC channel is connected. If the child process crashes in the interim, the usual OnProcessCrashed handler is invoked.

This is important for the GPU child process because it is prone to crashing before the IPC channel is established; it has to initialize Direct3D before the sandbox is turned on. This would lead to a browser hang if Direct3D or the driver crashed during initialization.

BUG=177611

Review URL: https://chromiumcodereview.appspot.com/12374061