RdpDesktopSession COM class should only use raw interfaces.
The daemon has to provide an implementation of IRdpDesktopSessionEventHandler interface in order to receive events from RdpDesktopSession. It is currently a dual interface (inherited from IDispatch) which either makes the implementation more complex than necessary or makes the daemon to depend on ATL. Switching to raw interfaces allows to both simplify the implementation and avoid ATL dependency.

This CL makes remoting_core.dll a proxy/stub DLL used to marshall the raw interfaces that the host defines.

BUG=137696

Review URL: https://codereview.chromium.org/12440003