ÜC: support DelegatedFrameData in Aura
Note: this tries to share as much as possible from the texture path,
unfortunately there's a bit of impedence mismatch between the
SwapBuffersPrepare/SwapBuffersCompleted steps and what needs to be done for
DelegatedFrameData. It will be easier to do if/when we remove the
TextureImageTransportSurface path.
    
BUG=146080,123444

Review URL: https://chromiumcodereview.appspot.com/11414092