Android: Support GPU process
This supports running the GPU thread in its own process outside the browser (which is really the default mode).

This changes the browser compositor API to pass a Java Surface instead of ANativeWindow. (Needed to pass the surface handle over to the other process).
It implements a synchronous API from the GPU thread to the browser to retrieve the surface, which is safe since this ends up on a binder thread different from the main thread.

This is enabled by removing kInProcessGpu from content_startup_flags.cc

Review URL: https://chromiumcodereview.appspot.com/12390077