New API in VideoCaptureDevice::EventHandler which allows shared-memory output buffers to be written directly by a VideoCaptureDevice.
Implement this API in VideoCaptureController. As part of the implementation, split all the DIB-management logic out of the VideoCaptureController, and into a new class with a well-defined interface, VideoCaptureBufferPool. In essence, VideoCaptureBufferPool gets all the code that previously lived behind the DIB lock in VideoCaptureController.

VideoCaptureBufferPool must be refcounted so as to extend the lifetime of the shared memory buffers until both the VideoCaptureController and VideoCaptureDevices are done with them.

TBR=jam@chromium.org
BUG=174520
TEST=content_unittests

Review URL: https://chromiumcodereview.appspot.com/12321092