Handle LibraryLoader error from the renderer process.
If we fail to load the library in the renderer process, just abort early.
This prevents an extra user-visible dialog that Chrome has crashed. Note
that we currently assume the browser process hits the same error and will
prompt the user. If this only happens for the renderer process we'd end up
with a hung tab, but it's not clear how that would happen.

As part of this, convert the RESULT_CODES enum to be generated from a template so that we can re-use the same values for C++ and Java

BUG=180054


Review URL: https://chromiumcodereview.appspot.com/12464004