Remove close_on_deactivate=false for tray bubbles (2nd)
Actually we want to close those bubbles on deactivation (like, opening
settings page or opening the chat window for a notification). To deal
with the race condition of close-on-deactivate and button action,
a static delay is introduced.

The first attempt (r184572) has been reverted due to a crash bug.
It seems that the crash happens because of missing cleanup of system
notification bubbles.

BUG=179992, 177075
TEST=on goobuntu, open the systme tray, wait for the power notifications, then click 'settings' and see no crash.


Review URL: https://chromiumcodereview.appspot.com/12510005