chromeos: Add DisplayPowerServiceProvider.
This makes Chrome export a SetDisplayPower D-Bus method call
for the power manager.

It also reworks OutputConfigurator to cache the
most-recently-requested power state and use it for future
display mode requests.

Finally, it works around a related bug where multiple mouse
events may be generated when the displays are reconfigured,
which would result in a report of user activity that could
abort suspending.

BUG=chromium-os:39289,180348,chrome-os-partner:12662
TBR=sky@chromium.org

Review URL: https://chromiumcodereview.appspot.com/12391004