Remove 'RequestFocus' for now, it doesn't make any effects.
The scroll view requests the focus, but it doesn't have any effects
on accessibility / text-to-speech right now but creates the border of
dotted lines.

The dotted line become visible as the result of crrev.com/12387067
because it is basically transparent, so we should remove the focus
for now to prevent this border.

The accessibility and tab focus has to be implemented, but it'll be done
in another CL.  See crbug.com/179431

BUG=179446


Review URL: https://chromiumcodereview.appspot.com/12528002