Allow renderer compositor message loop to be set by content embedder
Accomplished by adding a override method to ContentRendererClient
to return a MessageLoop, which if not null, is used as the renderer
compositor thread. In the process, make the ownership of the
compositor thread into RenderThreadImpl and rename CompositorThread
to InputHandlerManager.

Added a thread merge command line switch to control this in
android_webview.

BUG=179436


Review URL: https://chromiumcodereview.appspot.com/12383056