Allow impl-only animations, and return opacity values via AnimationEvents.
We would like to allow animations to run only on the impl thread, so we need to make sure they don't get clobbered when finished animations are removed. Also, we need a mechanism to allow opacity values to propagate back to the main-thread so that impl-only opacity animations have their values reflected on the main thread when the animation is complete.

BUG=165842


Review URL: https://chromiumcodereview.appspot.com/12453010