Fixed channel references in apps and extension intros
Now that Google Chrome 26 is stable, the intros need to reflect this clearly.

I also added a release note to extension's 'What's New' for pushMessaging API,
which went stable in M24.

Note that I did not add lots more information about apps-specific APIs to extension's 'What's New', as so far, this doc has been for extension APIs.

Working on a new way to document and promote changes across apps and extensions, so holding off on putting a lot of effort into the existing doc.

Review URL: https://codereview.chromium.org/12433018