Enable accelerated animations for orphan layers
This defines RenderWidgetCompositor::registerForAnimations,
which associates the LayerAnimationController of
newly-created cc::Layers with the AnimationRegistrar of the
RenderWidgetCompositor's LayerTreeHost. This also makes
cc::Layer::addAnimation accept animations whenever the
layer's LayerAnimationController has an AnimationRegistrar,
and makes LayerTreeHost only create an AnimationRegistrar
when accelerated animations are enabled.

Note that this is expected to regress Robohornet Pro by
about 16%: because of
https://bugs.webkit.org/show_bug.cgi?id=111617, Robohornet
Pro takes longer to complete when its animations are
accelerated. However, animations during this test (and on
other pages where animations are added to orphan layers)
will now be much smoother.

BUG=129683,179196

Review URL: https://chromiumcodereview.appspot.com/12388093