[sync] Upstream the Android ProfileSyncService
This creates a Java-layer ProfileSyncService that is the only class that
should be used from downstream Chrome for Android. The native
counterparts should not be used.

In the process, a hashing utility and also, some identity code and their
respective tests had to be upstreamed. No functional changes have been
made to them.

BUG=159203
TBR=sky@chromium.org

Review URL: https://codereview.chromium.org/12313075