[chromedriver] Fix SyncWebSocketImplTest.Reconnect in pyauto bots.
One purpose of this test is to check if the message queue of SyncWebSocket is cleared after reconnection.
The failure is caused by stopping the server right after pushing request message to SyncWebSocket.
The server has no time to response or even might not receive the request yet.

NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/12586002