OSX app list pagination and gesture scrolling.
This change adds pagination to the OSX app list by creating an 
additional layer of NSCollectionView. Each page has an NSCollectionView 
with up to 16 items, and the pages are stored in an enclosing 
NSCollectionView with a single row. This allows items to be laid out in 
clusters on each page to mimic the behavior on ChromeOS and Windows 
when scrolling between pages.

Gesture scrolling is enabled, and it snaps to the nearest page
when the gesture finishes.

BUG=138633
TEST=Added tests: AppsGridControllerTest.{Pagination, TwoPageModel}

Review URL: https://chromiumcodereview.appspot.com/12319090