Update PyAuto test to use updated barcode_decoder.py
The script has been improved so it's working with the
C++ version of Zxing instead of Java.
The C++ version of the tool doesn't need/accept barcode size so I removed all references to the barcode width/height in the test.

Because of this the barcode scripts has also moved into
webrtc/tools/barcode_tools so they will be checked out as a part of the normal
WebRTC checkout, which means the path must be updated in here.

Handling of return codes of the external tools used is also improved.

This CL should be committed after http://review.webrtc.org/1099007/ is committed.

BUG=none
TEST=successful local tests with Linux, Mac and Windows.
NOTRY=True

Review URL: https://chromiumcodereview.appspot.com/12493006