components: Rename components_tracing.gyp to follow the same pattern of the other gyp files.
We renamed it to tracing.gyp so it conforms with the other gyp files.

components_tracing_untrusted.gyp was renamed to tracing_untrusted.gyp as well. And left as a separate gyp file as it's built as untrusted nacl and has its own
special build requirements which may conflict with the other .gypi files.

BUG=138280
R=jbauman@chromium.org,joi@chromium.org
TBR=brettw@chromium.org

Review URL: https://chromiumcodereview.appspot.com/12599004