WebContentsVideoCaptureDevice: never drop frames.
Use VideoCaptureDevice::EventHandler::ReserveOutputBuffer() to abolish all cases of dropping half-captured frames in this code. This also gets us down to a minimal number of memory copies.

Because Deliver is very fast, it doesn't need its own thread or class anymore. We may deliver from the manager thread.

Render now writes directly to a VideoFrame, meaning it's now responsible for colorspace conversion. I've rewritten Render with this in mind, and also, I use some recently-added video_util functions.

The CaptureMachine is changed so that we cap the number of outstanding frames at various stages of the pipeline. The limits set here are 1 capture at a time, and 2 frames total in (Capture + Render + Deliver). Frames that are downstream of Deliver aren't explicitly tracked by the machine, but the BufferPool will exhaust in that case.

When the CaptureMachine limits are hit, the Machine enters the "capture_suspended_" state, meaning that the next capture will be started upon completion of the current one.

The unit tests are changed accordingly.

BUG=155174,174543
TEST=content_unittests, manual test using selfmirroring tab capturing extension

Review URL: https://chromiumcodereview.appspot.com/12417002