History: Fix showing the search results twice.
Initially setSearch received two parameters, one optional. After the recent
refactoring not receving the second parameter would call setPageState with
value 'undefined' for that parameter which would lead to a second call to
setPageState with the value 0 which showed the results twice.

This CL deletes that parameter as was not needed anymore.

R=dubroy@chromium.org
BUG=180952

Review URL: https://chromiumcodereview.appspot.com/12635003