Really fix crash when dragging and dropping in chrome://settings/startup
The root cause of this crash was that dropping in the <input> field
dispatches a DragDropStartupPage event to the C++ handler. Since the
<input> field doesn't have a corresponding index in the C++ model,
CustomHomePagesModel would attempt to dereference an index past the end
of |entries_|. The fix is to suppress drop events on the <input> field
from triggering the C++ handler, since you can't move the <input> field
anyway.

Also clean up the code to use ints (and remove conversions to/from
strings) where appropriate.

BUG=132974
TBR=jhawkins

Review URL: https://chromiumcodereview.appspot.com/12548008