Remove the action choose dialog's icon and title.
By creating a browser and later adding a tab using chrome::AddSelectedTabWithURL it is impossible to get rid of the icon and the title.
However, chrome::OpenApplication can do that, but it doesn't allow to pass dimensions. Therefore, an argument has been added to be able to override bounds.

TEST=Insert an sd card with a DCIM folder. Verify the action choose dialog.
BUG=179084

Review URL: https://chromiumcodereview.appspot.com/12491008