Adds support of local audio streams for more than one media player.
Main goal is to allow the user to use a local stream as source for more than one audio element. Adding the stream to one extra element will result in an increased output volume and the possibility to tune the volume on each element. If two elements are used, both elements must be muted to silence the audio.

The design is now more clear since more work is done in the local renderer and not divided between the capturer (which used to own the FIFO) and the local renderer. The capturer only dispatches the recorded audio out to its sinks and they deal with the rest.

- Removes auto-mute on local media streams when rendered by a media player.
- Moves local FIFO from WebRtcAudioCapturer to WebRtcLocalAudioRenderer.
- Now allows multiple WebRtcLocalAudioRenderer objects where each instance creates its own output stream.
- Adds support for the enabled attribute on the MediaStreamTrack object.

BUG=none
TEST=All existing WebRTC demos and a new test at: https://www.corp.google.com/~henrika/webrtc/multi_local_rendering.html
TBR=wjia@chromium.org

Review URL: https://codereview.chromium.org/12599002