Check that DB initialized properly in PerformanceMonitor
Check that each LevelDB instance will be correctly opened prior to any use in
PerformanceMonitor.  If one does not open, try to fix it.  If it still does
not open, do not finish PerformanceMonitor's initialization process (so that
nothing will use that DB, and we won't crash).

BUG=168208


Review URL: https://chromiumcodereview.appspot.com/12385056