Avoid sending nonempty view size with empty backing size.
This condition, which occurs during EmbedContentViewActivity
initialization, results in a white-screen state because of the following
sequence:
1. pending_resize_ack_ is set in RWHI.
2. didCompleteSwapBuffers is never called in RenderWidget because of the
empty backing, therefore the resize ack is never sent.
3. When the backing size is set to a nonzero value, RWHI refuses to send
it because it expects a pending resize.

We can avoid this situation by returning 0 from GetViewBounds as well
until the backing is initialized.

BUG=181244


Review URL: https://chromiumcodereview.appspot.com/12430010