Add reviewers to optional outputs for changes
This adds a comma-separated list of reviewers to the optional outputs
for changes. This can be added to the output using the custom formatter
option.  E.g. --output-format-changes="{url} {title} <{reviewers}>"

Also, for webkit changes, use CommitterList to map reviewer names to
emails so that they are similar format to other changes.


Review URL: https://chromiumcodereview.appspot.com/12446008