cc: Prevent crash when context creation fails.
If context creation fails, don't store the context provider. Otherwise,
when WebKit asks for the GrContext, we create it with a WebGraphicsContext3D
which cannot be made current, and this causes a crash in skia.  We already
know that the Context3D is no good, because BindToCurrentThread() on the
provider failed. So we can avoid giving skia the Context3D and having it
crash.

BUG=180911

Review URL: https://chromiumcodereview.appspot.com/12771011