Fix Notifications Galore test app settings bug.
Fixes a bug that caused Galore to not work when the latest version was
run for the first time or when it was uninstalled & reinstalled. The bug
caused the app to fail to start up when no settings for it were found in
synced storage. This fixes that.

BUG=NONE
TBR=miket@chromium.org

Review URL: https://codereview.chromium.org/12492004