Fix empty candidate window for IME Extension API.
The problem is that some engines are never deactivated when the engine is switched.
The original IBus spec assumes each engine instance will be created on CreateEngine method and destroyed on Destroy function.
However in the case of Chrome, we switch the engine instance at the next CreateEngine function and no-op for Destroy function.
This is ideally okay because there is only one engine instance.
However ibus-daemon sends "Disable" message *AFTER* CreateEngine function, so Extension IME missed it, this is the root cause.

I changed to calling "Disable" function at the instance switching and it fixes this problem.

BUG=182916
TEST=Manually checked on lumpy.


Review URL: https://chromiumcodereview.appspot.com/12598004