Change Chrome OS file chooser private API toward enabling Drive everywhere.
This patch
- changes the query parameter name "disableDrive" to "returnLocalPath",
 since we are going to enable Drive everywhere, by changing
 behavior depending on whether the caller supports Drive path or not.
- adds the parameters "needLocalPath" and "forOpen" to the
fileBrowserPriavate.selectFile API. Open and Save dialog needs
different kind of path resolution.

This CL just changes the interface. Actual use of it will come in later patches.
For the detail of the rationale and future plan of how to
use the changed APIs, see the bug tracker.

BUG=181494
TEST=Open/Save dialog on ChromeOS just works. No behavior change.

Review URL: https://codereview.chromium.org/12641006