JNI on Android: refresh self-documenting "SampleForTests.java".
This file is the authoritative source of documentation and information on how to
generate JNI bindings.
However, it was probably too specific and probably required a fair bit of previous knowledge.
This is an attempt in making it more comprehensive, aiming to have these files as a full guide
for all supported cases and the best practices, and help everyone work with JNI more effectively.

BUG=
TEST=jni_generator_tests


Review URL: https://chromiumcodereview.appspot.com/12600014