Add detection to FeatureInfo to see if samplers are available.
Not all platforms have Samplers available. We detect if we're running
as OpenGL ES 3.x or if we have GL_ARB_sampler_objects available. If
so we set enable_samplers in the feature info object.

This will be extended in a future CL to advertise CHROMIUM_sampler_objects
as an extension, and used to determine if the implementations should
fail with INVALID_OPERATION.

BUG=179849
R=gman@chromium.org


Review URL: https://chromiumcodereview.appspot.com/12841003