android: Ignore the first scroll delta to avoid a jump
Ignore the very first scroll delta we get from the gesture detector to
avoid a visible jump at the start of scrolling. This is needed because
the gesture detector has a dead zone that inhibits detecting small
movements of the finger as scrolling. Once the finger exists the
dead zone, the very first scroll delta also includes the delta between
where the finger touched the glass and where it exited the dead zone. If
this delta is used to scroll, then the page will always jump by the size
of the dead zone at the start of a scroll.

BUG=181521


Review URL: https://chromiumcodereview.appspot.com/12691005