[Instant] Keep the previous default provider suggestion if input is similar
Previously was discarding the |default_provider_suggestion_| in all cases. Now checking for staleness of the previous suggestion before discarding, since under similar cases, the same suggestion is likely to come back. This avoids flickering in the omnibox as we discard the suggestion and get the exact same one.

BUG=159661, 166084
TEST=under Instant extended, type [nyti] and notice no flickering when completing to [mes.com] as input is contained in the suggestion (and we are not discarding it each time).


Review URL: https://chromiumcodereview.appspot.com/12330182