Remove close_on_deactivate=false for tray bubbles (3rd)
Actually we want to close those bubbles on deactivation (like, opening
settings page or opening the chat window for a notification).

close_on_deactivate=true will close the bubble when mouse-pressed,
but the next mouse-released event on the tray_ will invoke
PerformAction method to reopen the bubble. To prevent this reopen,
the mouse capture of the tray has to be released.

BUG=177075,179992


Review URL: https://chromiumcodereview.appspot.com/12701021