Build a new TokenCacheService so I can stop using TokenService.
When building code to cache the Obfuscated Gaia Id that we receive from the server, we had used the TokenService to cache it.  However, the Token service wasn't designed to cache arbitrary tokens, it was designed to fetch tokens, and cached the fetched token as a side effect.

This patch delivers a new TokenCache service that does nothing more than cache tokens on a per-profile basis.  The Cache will be cleared out when the user signs out and when chromium exits.

This patch also removes the code I added to TokenService to allow manually adding a token, since we no longer need that code.


BUG=177125

Review URL: https://chromiumcodereview.appspot.com/12380006