Implement a policy to autologin a public account.
This introduces a policy to automatically login a public account on
Chrome OS after a specified period of time has elapsed at the login
screen.  There are two new policy settings:
"DeviceLocalAccountAutoLoginUsername", which determines which public
account will be logged in, and "DeviceLocalAccountAutoLoginTimerMillis",
which specifies the amount of time that should elapse before autologin
takes place.

The autologin timer is started in ExistingUserController when either the
sign-in screen UI is finished loading or the aforementioned policy
settings are changed.

BUG=152933


Review URL: https://chromiumcodereview.appspot.com/12218078