Connect RenderWidgetHostViewAura to BrowserAccessibilityManager.
This change doesn't completely make Aura/Win accessible, it just
refactors BrowserAccessibilityManager a bit so that it can be
used by both Win and AuraWin without convoluted typecasting.

Specifically, the assumption before was that
BrowserAccessibilityManager could be initialized with a
parent gfx::NativeView that's the accessibility parent.
That abstraction doesn't work in Aura - the
parent accessible object is not the same as the parent HWND
(and BrowserAccessibilityManagerWin needs both). So rather than
trying to have the parent accessible object be a
cross-platform concept, I just made it a platform-specific
concept that the base class doesn't know about now.

BUG=175156


Review URL: https://chromiumcodereview.appspot.com/12555004