UI updates for items in the OSX app list: mouseover, font, padding.
This change adds a CrTrackingArea to items in the apps grid for a
mouseover effect. There are also minor UI tweaks: use a correct padding
for the icon in the apps grid, and use the same font as used by the
CrOS and Windows app launchers.

BUG=138633
TEST=Added: app_list_unittests --gtest_filter=AppsGridControllerTest.MouseoverSelects

Review URL: https://chromiumcodereview.appspot.com/12764024